pyocr_to_api/requirements.txt
boto==2.42.0
bz2file==0.98
cffi==1.8.3
decorator==4.0.10
gensim==0.13.2
img2pdf==0.2.1
ipython==5.1.0
ipython-genutils==0.1.0
numpy==1.11.1
pandas==0.18.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.3.1
prompt-toolkit==1.0.7
ptyprocess==0.5.1
pycparser==2.14
Pygments==2.1.3
pyocr==0.4.1
PyPDF2==1.26.0
pytesseract==0.1.6
python-dateutil==2.5.3
pytz==2016.6.1
reportlab==3.3.0
requests==2.11.1
ruffus==2.6.3
scipy==0.18.1
simplegeneric==0.8.1
six==1.10.0
smart-open==1.3.4
traitlets==4.3.0
Wand==0.4.3
wcwidth==0.1.7
pytest==3.0.2
